Separated toolbar from topology panel
